Add plumbing for port references
- Also moved some of the files around for clarity and correct the name of 'Well Known Ports'
